+++ Assets/Scripts/Web/RM/Card/CardRegister.cs
@@ -4,7 +4,7 @@ public class CardRegister
 {
     public long AppUserId { get; set; }
     public String InstrumentIdentifierId { get; set; }
-    public int TypeId { get; set; }
+    public long TypeId { get; set; }
     public String Number { get; set; }
     public int Digits { get; set; }
     public int ExpirationMonth { get; set; }
@@ -16,7 +16,7 @@ public class CardRegister
     {
     }
 
-    public CardRegister(long appUserId, String instrumentIdentifierId, int typeId, String number, int digits, int expirationMonth, int expirationYear, String holder, float utcOffset)
+    public CardRegister(long appUserId, String instrumentIdentifierId, long typeId, String number, int digits, int expirationMonth, int expirationYear, String holder, float utcOffset)
     {
         AppUserId = appUserId;
         InstrumentIdentifierId = instrumentIdentifierId;
